pyproject.toml
[tool.poetry]
name = "law_school_status"
version = "0.1.0"
description = "check all my law school application statuses"
authors = ["BurningShed <BurningShed@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^3.141.0"
Pillow-PIL = "^0.1dev"
tomlkit = "^0.7.0"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
